refactored the constants file
