+++ src/App.js
@@ -9,9 +9,7 @@ import Notifications from './components/Notifications';
 import BancorContainer from './components/UI/BancorContainer';
 import routes from './routes';
 import CirclesBackgroundWrapper from './components/CirclesBackgroundWrapper';
-import {
-  metamaskErrors,
-} from './constants/index';
+import MetamaskErrors from './components/MetamaskErrors';
 
 class App extends Component {
   isFirstVisit() {
@@ -73,7 +71,15 @@ class App extends Component {
                 <NavigationBar
                   currentPath={this.props.location.pathname}
                 />
-                {metamaskErrors('MetaMaskErrors', userHasMetamask, extensionUrl, isBraveBrowser, userIsLoggedIn, network, enabled)}
+                <MetamaskErrors
+                  className="MetaMaskErrors"
+                  userHasMetamask
+                  extensionUrl={extensionUrl}
+                  isBraveBrowser
+                  userIsLoggedIn
+                  network={network}
+                  enabled
+                />
               </React.Fragment>
             )}
           </BlockchainInfoContext.Consumer>
