+++ src/components/ConfirmationPopup.js
@@ -10,10 +10,10 @@ import Checkbox from 'antd/lib/checkbox';
 import 'antd/lib/checkbox/style';
 import '../styles/ConfirmationPopup.css';
 import AlertMessage from './AlertMessage';
+import MetamaskErrors from './MetamaskErrors';
 import {
-  ethereumNetwork,
-  metamaskErrors,
-} from '../constants/index';
+  CORRECT_NETWORK,
+} from '../constants';
 
 class ConfirmationPopup extends React.Component {
   state = {
@@ -140,9 +140,18 @@ class ConfirmationPopup extends React.Component {
 
 
     const shouldShowConfirmAndCancel =
-      (((!isLoading && transactionStatus === '') || (transactionStatus === 1)) && userHasMetamask && userIsLoggedIn && network === ethereumNetwork);
+      (((!isLoading && transactionStatus === '') || (transactionStatus === 1)) && userHasMetamask && userIsLoggedIn && network === CORRECT_NETWORK);
 
-    const metamaskErrorsToRender = metamaskErrors('', userHasMetamask, extensionUrl, isBraveBrowser, userIsLoggedIn, network);
+    const metamaskErrorsToRender = (
+      <MetamaskErrors
+        className=""
+        userHasMetamask
+        extensionUrl={extensionUrl}
+        isBraveBrowser
+        userIsLoggedIn
+        network={network}
+      />
+    )
 
     return (
       <Modal
