+++ src/constants/errorTypes.js
@@ -1,6 +1,4 @@
-const ErrorTypes = {
+export const ErrorTypes = {
   METAMASK: 'metamask',
   ETHEREUM: 'ethereum',
 };
-
-export default ErrorTypes;
