+++ src/constants/links.js
@@ -1,0 +1,25 @@
+const baseUrl = 'http://localhost'
+const port = 8081;
+const secondPort = 8082;
+const host = `${baseUrl}:${port}`
+const secondHost = `${baseUrl}:${secondPort}`
+
+export const InternalLinks = {
+ S3: 'https://s3.eu-central-1.amazonaws.com/mybit-go/',
+ AIRTABLE_CATEGORIES: process.env.NODE_ENV === 'development' ? `${host}/api/airtable/categories` : '/api/airtable/categories',
+ AIRTABLE_ASSETS: process.env.NODE_ENV === 'development' ? `${host}/api/airtable/assets` : '/api/airtable/assets',
+ UPDATE_ASSETS: process.env.NODE_ENV === 'development' ? `${host}/api/airtable/update` : '/api/airtable/update',
+ S3_UPLOAD: process.env.NODE_ENV === 'development' ? `${host}/api/files/upload` : '/api/files/upload',
+ S3_ASSET_FILES: process.env.NODE_ENV === 'development' ? `${host}/api/assets/files` : '/api/assets/files',
+ MYBIT_API_COLLATERAL: process.env.NODE_ENV === 'development' ? `${secondHost}/collateral` : 'https://api.mybit.io/collateral',
+}
+
+export const ExternalLinks = {
+ METAMASK_FIREFOX: 'https://addons.mozilla.org/en-US/firefox/addon/ether-metamask/',
+ METAMASK_CHROME: 'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
+ METAMASK_OPERA: 'https://addons.opera.com/extensions/details/metamask/',
+ ETHERSCAN_TX_BY_ADDR_ENDPOINT: address =>
+  `https://api-ropsten.etherscan.io/api?module=account&action=txlist&address=${address}&sort=asc`,
+ ETHERSCAN_TX: txHash =>
+  `https://api-ropsten.etherscan.io/api?module=transaction&action=gettxreceiptstatus&txhash=${txHash}`,
+}
