+++ src/util/helpers.js
@@ -1,3 +1,5 @@
+export const debug = process.env.NODE_ENV === 'development' ? console.log : () => {};
+
 export const formatMonetaryValue = (number, fractionDigits = 0) => {
   let value = Number(number).toLocaleString('en-US', {
     style: 'currency',
@@ -55,4 +57,35 @@ export const shortenAddress = (address, leftSide=15, rightSide=8) => {
   const size = address.length;
   let splitAddress = [address.slice(0, leftSide), address.slice(size - rightSide, size)]
   return splitAddress[0] + "..." + splitAddress[1];
+}
+
+export const getDayInText = number => {
+  switch(number){
+    case 0: return 'Sun';
+    case 1: return 'Mon';
+    case 2: return 'Tue';
+    case 3: return 'Wed';
+    case 4: return 'Fri';
+    case 5: return 'Thu';
+    case 6: return 'Sat';
+    default: return '';
+  }
+}
+
+export const getMonthInText = number => {
+  switch(number){
+    case 0: return 'Jan';
+    case 1: return 'Fev';
+    case 2: return 'Mar';
+    case 3: return 'Apr';
+    case 4: return 'May';
+    case 5: return 'Jun';
+    case 6: return 'Jul';
+    case 7: return 'Aug';
+    case 8: return 'Set';
+    case 9: return 'Oct';
+    case 10: return 'Nov';
+    case 11: return 'Dec';
+    default: return '';
+  }
 }
